Extract runLinter function to separate file to clean up index.ts
