+++ packages/perseus-linter/src/types.ts
@@ -1,3 +1,7 @@
+/**
+ * The shape of the linter context object that is passed through the
+ * tree with additional information about what we are checking.
+ */
 export type LinterContextProps = {
     contentType: string;
     highlightLint: boolean;
